Move eloquent model methods into base model, create support JsonEncodingException
Backward incompatible change: When casting an attribute to json, if encoding fails, a Support\JsonEncodingException is thrown rather than a Database\Eloquent\JsonEncodingException
