+++ src/Illuminate/Database/Eloquent/JsonEncodingException.php
@@ -2,9 +2,9 @@
 
 namespace Illuminate\Database\Eloquent;
 
-use RuntimeException;
+use Illuminate\Support\JsonEncodingException as BaseJsonEncodingException;
 
-class JsonEncodingException extends RuntimeException
+class JsonEncodingException extends BaseJsonEncodingException
 {
     /**
      * Create a new JSON encoding exception for the model.
@@ -16,20 +16,5 @@ class JsonEncodingException extends RuntimeException
     public static function forModel($model, $message)
     {
         return new static('Error encoding model ['.get_class($model).'] with ID ['.$model->getKey().'] to JSON: '.$message);
-    }
-
-    /**
-     * Create a new JSON encoding exception for an attribute.
-     *
-     * @param  mixed  $model
-     * @param  mixed  $key
-     * @param  string $message
-     * @return static
-     */
-    public static function forAttribute($model, $key, $message)
-    {
-        $class = get_class($model);
-
-        return new static("Unable to encode attribute [{$key}] for model [{$class}] to JSON: {$message}.");
     }
 }
