+++ src/Illuminate/Database/Eloquent/Model.php
@@ -143,9 +143,7 @@ abstract class Model extends BaseModel implements QueueableEntity, UrlRoutable
     {
         $this->bootIfNotBooted();
 
-        $this->syncOriginal();
-
-        $this->fill($attributes);
+        parent::__construct($attributes);
     }
 
     /**
@@ -205,7 +203,7 @@ abstract class Model extends BaseModel implements QueueableEntity, UrlRoutable
     }
 
     /**
-     * Get an attribute from the model.
+     * Get an attribute or relation from the model.
      *
      * @param  string  $key
      * @return mixed
@@ -992,36 +990,7 @@ abstract class Model extends BaseModel implements QueueableEntity, UrlRoutable
      */
     public function toArray()
     {
-        return array_merge($this->attributesToArray(), $this->relationsToArray());
-    }
-
-    /**
-     * Convert the model instance to JSON.
-     *
-     * @param  int  $options
-     * @return string
-     *
-     * @throws \Illuminate\Database\Eloquent\JsonEncodingException
-     */
-    public function toJson($options = 0)
-    {
-        $json = json_encode($this->jsonSerialize(), $options);
-
-        if (JSON_ERROR_NONE !== json_last_error()) {
-            throw JsonEncodingException::forModel($this, json_last_error_msg());
-        }
-
-        return $json;
-    }
-
-    /**
-     * Convert the object into something JSON serializable.
-     *
-     * @return array
-     */
-    public function jsonSerialize()
-    {
-        return $this->toArray();
+        return array_merge(parent::toArray(), $this->relationsToArray());
     }
 
     /**
@@ -1420,60 +1389,11 @@ abstract class Model extends BaseModel implements QueueableEntity, UrlRoutable
     }
 
     /**
-     * Dynamically retrieve attributes on the model.
-     *
-     * @param  string  $key
-     * @return mixed
+     * @throws \Illuminate\Database\Eloquent\JsonEncodingException
      */
-    public function __get($key)
+    protected function throwJsonEncodingException()
     {
-        return $this->getAttribute($key);
-    }
-
-    /**
-     * Dynamically set attributes on the model.
-     *
-     * @param  string  $key
-     * @param  mixed  $value
-     * @return void
-     */
-    public function __set($key, $value)
-    {
-        $this->setAttribute($key, $value);
-    }
-
-    /**
-     * Determine if the given attribute exists.
-     *
-     * @param  mixed  $offset
-     * @return bool
-     */
-    public function offsetExists($offset)
-    {
-        return ! is_null($this->getAttribute($offset));
-    }
-
-    /**
-     * Get the value for a given offset.
-     *
-     * @param  mixed  $offset
-     * @return mixed
-     */
-    public function offsetGet($offset)
-    {
-        return $this->getAttribute($offset);
-    }
-
-    /**
-     * Set the value for a given offset.
-     *
-     * @param  mixed  $offset
-     * @param  mixed  $value
-     * @return void
-     */
-    public function offsetSet($offset, $value)
-    {
-        $this->setAttribute($offset, $value);
+        throw JsonEncodingException::forModel($this, json_last_error_msg());
     }
 
     /**
@@ -1484,29 +1404,9 @@ abstract class Model extends BaseModel implements QueueableEntity, UrlRoutable
      */
     public function offsetUnset($offset)
     {
-        unset($this->attributes[$offset], $this->relations[$offset]);
-    }
+        parent::offsetUnset($offset);
 
-    /**
-     * Determine if an attribute or relation exists on the model.
-     *
-     * @param  string  $key
-     * @return bool
-     */
-    public function __isset($key)
-    {
-        return $this->offsetExists($key);
-    }
-
-    /**
-     * Unset an attribute on the model.
-     *
-     * @param  string  $key
-     * @return void
-     */
-    public function __unset($key)
-    {
-        $this->offsetUnset($key);
+        unset($this->relations[$offset]);
     }
 
     /**
@@ -1535,16 +1435,6 @@ abstract class Model extends BaseModel implements QueueableEntity, UrlRoutable
     public static function __callStatic($method, $parameters)
     {
         return (new static)->$method(...$parameters);
-    }
-
-    /**
-     * Convert the model to its string representation.
-     *
-     * @return string
-     */
-    public function __toString()
-    {
-        return $this->toJson();
     }
 
     /**
